+++ src/OpenFeature.Providers.Flagd/Resolver/InProcess/JsonEvaluator.cs
@@ -230,8 +230,11 @@ internal class JsonEvaluator
         {
             if ("DISABLED" == flagConfiguration.State)
             {
-                throw new FeatureProviderException(ErrorType.FlagNotFound,
-                    "FLAG_NOT_FOUND: flag '" + flagKey + "' is disabled");
+                return new ResolutionDetails<T>(
+                    flagKey: flagKey,
+                    defaultValue,
+                    reason: Reason.Disabled
+                );
             }
 
             Dictionary<string, object> combinedMetadata = _flagSetMetadata.ToDictionary(
@@ -336,7 +339,8 @@ internal class JsonEvaluator
 
                 // if variant is null, revert to default
                 reason = Reason.Default;
-                flagConfiguration.Variants.TryGetValue(flagConfiguration.DefaultVariant,
+                variant = flagConfiguration.DefaultVariant;
+                flagConfiguration.Variants.TryGetValue(variant,
                     out var defaultVariantValue);
                 if (defaultVariantValue.ValueKind == JsonValueKind.Undefined || defaultVariantValue.ValueKind == JsonValueKind.Null)
                 {